Move some functions to new file
