+++ custom/zipline/tasks.py
@@ -1,8 +1,8 @@
 import json
 import requests
 from celery.task import task
+from custom.zipline.api import ProductQuantity, get_order_update_critical_section_key
 from custom.zipline.models import EmergencyOrder, EmergencyOrderStatusUpdate
-from datetime import datetime
 from django.conf import settings
 from dimagi.utils.couch import CriticalSection
 from dimagi.utils.logging import notify_exception
@@ -20,41 +20,6 @@ REQUEST_TIMEOUT = 120
 # See send_emergency_order_request()
 RETRY_INTERVAL = 5
 MAX_ATTEMPTS = 3
-
-
-class ProductQuantity(object):
-    """
-    A simple class to describe a product and the quantity of it to be ordered.
-    """
-
-    def __init__(self, code, quantity):
-        """
-        :param code: the product code of the product being requested
-        :param quantity: the quantity of the product being requested
-        """
-        self.code = code
-        self.quantity = quantity
-
-
-def initiate_emergency_order(domain, user, phone_number, location, products):
-    """
-    :param domain: the domain in which the order is being requested
-    :param user: the user who is initiating the emergency order request
-    :param phone_number: the phone_number (string) of the user who is initiating the emergency order request
-    :param location: the SQLLocation where the products should be delivered
-    :param products: a list of ProductQuantity objects representing the products to be ordered
-    """
-    order = EmergencyOrder.objects.create(
-        domain=domain,
-        requesting_user_id=user.get_id,
-        requesting_phone_number=phone_number,
-        location=location,
-        location_code=location.site_code,
-        products_requested=[{'code': p.code, 'quantity': p.quantity} for p in products],
-        timestamp=datetime.utcnow()
-    )
-
-    send_emergency_order_request(order.pk)
 
 
 @task(ignore_result=True)
@@ -78,15 +43,6 @@ def send_emergency_order_request(order_id, attempt=1):
         )
         create_error_record(order, 'Internal error: {}'.format(str(e)))
         handle_emergency_order_request_retry(order, attempt)
-
-
-def get_order_update_critical_section_key(order_id):
-    """
-    :param order_id: the pk of the order
-    :return: the key to be used with CriticalSection for preventing multiple
-    threads from updating the same order at the same time
-    """
-    return 'zipline-updating-order-id-{}'.format(order_id)
 
 
 def _send_emergency_order_request(order, attempt):
